Extract `toc/locks.rs`
